requirements.txt
clang>=16.0.0
pyyaml>=6.0
